chore: clippy and idiomatic-Rust cleanup across pt-core, pt-common, pt-config
Systematic pass addressing Clippy lints and Rust idiom improvements across
57 files (291 insertions, 240 deletions). No behavioral changes except one
missing assertion added in action_real.rs.

Source code (26 files):
- Replace mutable-then-assign patterns with struct literals + ..Default::default()
  (clippy::field_reassign_with_default) across config, inference, calibrate,
  collect, decision, supervision, daemon, output, and session modules
- Fix clippy::needless_return: remove explicit return Ok(()) at function ends
- Fix clippy::get_first: replace .get(0) with .first()
- Fix clippy::option_map_flatten: replace .map().flatten() with .and_then()

Test files (23 files):
- Same mutable-default-then-assign → struct literal cleanup
- Replace ProcessHarness::default() with unit struct construction ProcessHarness
- Add missing assertion in action_real.rs zombie process test

Cargo.toml:
- Gate tempfile dependency behind dep: syntax for test-tempdir feature (hygiene fix)

+++ crates/pt-core/src/session/diff_tests.rs
@@ -291,7 +291,7 @@ mod tests {
                     i,
                     &format!("b1:{}:{}", i, i),
                     "useful",
-                    10 + (i % 30) as u32,
+                    10 + (i % 30),
                     "keep",
                     0.05,
                 )
@@ -311,7 +311,7 @@ mod tests {
                     i,
                     &format!("b1:{}:{}", i, i),
                     "useful",
-                    10 + (i % 30) as u32,
+                    10 + (i % 30),
                     "keep",
                     0.05,
                 )
